Updating paths
