+++ examples/teams/index.ts
@@ -1,4 +1,5 @@
-export { TeamPlatform } from './team-platform'
-export { TeamBurnham } from './team-burnham'
-export { TeamRiker } from './team-riker'
 export { TeamTroi } from './team-troi'
+export { TeamRiker } from './team-riker'
+export { TeamBurnham } from './team-burnham'
+export { TeamPlatform } from './team-platform'
+
