+++ examples/teams/team-riker/index.ts
@@ -1,4 +1,4 @@
-import { ClusterInfo } from '../../../lib/stacks/eks-blueprint-stack';
+import { ClusterInfo } from '../../../lib/stacks/cluster-types';
 import { Team } from '../../../lib/teams';
 
 export class TeamRiker implements Team {
