+++ examples/teams/team-troi/index.ts
@@ -2,8 +2,8 @@ import * as cdk from '@aws-cdk/core';
 import * as eks from "@aws-cdk/aws-eks";
 import * as s3 from "@aws-cdk/aws-s3";
 
-import { ClusterInfo } from '../../../lib/stacks/eks-blueprint-stack';
 import { Team } from '../../../lib/teams';
+import { ClusterInfo } from '../../../lib/stacks/cluster-types';
 
 export class TeamTroi implements Team {
     readonly name: string = 'team-troi';
