+++ lib/addons/argocd/index.ts
@@ -1,8 +1,8 @@
-import { ClusterAddOn, ClusterInfo } from "../../stacks/eks-blueprint-stack";
+import { ClusterAddOn, ClusterInfo } from "../../stacks/cluster-types";
 
 export class ArgoCDAddOn implements ClusterAddOn {
 
-    readonly namespace : string;
+    readonly namespace: string;
 
     constructor(namespace?: string) {
         this.namespace = namespace ?? "argocd";
