+++ lib/index.ts
@@ -1,8 +1,8 @@
-export * from './addons';
-export * from './cluster-providers'
-export * from './stacks';
-export * from './teams';
-
+export * from './stacks/cluster-types';
+export * as AddOns from './addons';
+export * as Stacks from './stacks';
+export * as Teams from './teams';
+export * as ClusterProviders from './cluster-providers'
